requirements.txt
--extra-index-url=https://databricks-dash.aws.plotly.host/packages
dash-design-kit==1.8.1
dash==2.9.3
gunicorn==20.0.4
pandas>=1.1.5

numpy
scikit-learn
celery
redis
dash-bootstrap-components
dash_ag_grid
dash-extensions==0.0.65
dash_pro_components-0.3.1.tar.gz
xgboost
dash_cool_components
